Relax ledger block decoding assertion
There's no need to assert if rlp decoding a block in the local ledger
fails. Ledger-tail operators have reported this failing sporadically in
practice. In addition, without this change, there's a potential DOS
vector where a malicious node could send blocksyncs for the network tip
in an attempt to trip this assertion.